refactored data_reader_jag_conduit into two classes:   base class: data_reader_conduit   derived class: data_reader_jag_conduit The derived class contains those functions that depend on conduit data types:   using ch_t = float;   using conduit_ch_t = conduit::float32_array;   using scalar_t = double;   using input_t = double; The reason is that the data types for hydra and JAG data may differ (they currently differ for ch_t and conduit_ch_t, which are double for the hydra data) A next step is to convert data_reader_jag_conduit into a templated class, where the above "using" statements become template parameters.
