+++ src/data_readers/CMakeLists.txt
@@ -21,6 +21,7 @@ set_full_path(THIS_DIR_SOURCES
   data_reader_imagenet_patches.cpp
   data_reader_jag.cpp
   data_reader_jag_conduit.cpp
+  data_reader_conduit.cpp
   data_reader_merge_features.cpp
   data_reader_merge_samples.cpp
   data_reader_mesh.cpp
